#381: move ExtraPath class to its own file
