+++ src/main/python/test_support/extrapath.py
@@ -1,0 +1,24 @@
+#
+# This file is licensed under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import sys
+
+class ExtraPath(object):
+    def __init__(self, extra_path):
+        self.extra_path = extra_path
+
+    def __enter__(self):
+        sys.path.insert(1, self.extra_path)
+
+    def __exit__(self, type, value, traceback):
+        sys.path.remove(self.extra_path)
